Replaced int with INT where possible in kernel/util.
git-svn-id: https://euler.mathematik.tu-chemnitz.de/repos/nfft/lib/branches/nfft-3.2@3860 c8b3e4ef-9df4-0310-9d76-b6dd82e1323a

+++ kernel/util/thread.c
@@ -24,13 +24,13 @@
 #include <omp.h>
 #endif
 
-int X(get_num_threads)(void)
+INT X(get_num_threads)(void)
 {
 #ifdef _OPENMP
-  int nthreads;
+  INT nthreads;
   #pragma omp parallel default(shared)
   {
-    int n = omp_get_num_threads();
+    INT n = (INT)omp_get_num_threads();
     #pragma omp master
     {
       nthreads = n;
